batch upload service (by triggering start_pipeline CLoud Run service)

+++ common/src/common/utils/copy_gcs_documents.py
@@ -21,7 +21,8 @@ from common.config import STATUS_IN_PROGRESS, STATUS_SUCCESS, STATUS_ERROR
 
 # disabling for linting to pass for blob_copy variable
 # pylint: disable = unused-variable
-def copy_blob(bucket_name, source_blob_name, destination_blob_name, dest_bucket_name=None):
+def copy_blob(bucket_name, source_blob_name, destination_blob_name,
+    dest_bucket_name=None, delete_original=False):
   storage_client = storage.Client()
   source_bucket = storage_client.bucket(bucket_name)
   source_blob = source_bucket.blob(source_blob_name)
@@ -31,5 +32,13 @@ def copy_blob(bucket_name, source_blob_name, destination_blob_name, dest_bucket_
     destination_bucket = storage_client.bucket(dest_bucket_name)
   blob_copy = source_bucket.copy_blob(source_blob, destination_bucket,
                                       destination_blob_name)
-  source_bucket.delete_blob(source_blob_name)
+  if delete_original:
+    source_bucket.delete_blob(source_blob_name)
+  return STATUS_SUCCESS
+
+
+def move_blob(bucket_name, source_blob_name, destination_blob_name,
+    dest_bucket_name=None):
+  copy_blob(bucket_name, source_blob_name, destination_blob_name,
+            dest_bucket_name=dest_bucket_name, delete_original=True)
   return STATUS_SUCCESS
